+++ sentinel/package.json
@@ -12,6 +12,7 @@
     "@medic/message-utils": "^1.0.2",
     "@medic/phone-number": "^1.0.0",
     "@medic/registration-utils": "^1.1.0",
+    "@medic/sentinel": "file:../../medic-webapp-sandbox/shared-libs/sentinel",
     "@medic/server-checks": "^1.0.1",
     "@medic/task-utils": "^1.0.0",
     "@medic/tombstone-utils": "^1.0.0",
