+++ sentinel/src/config.js
@@ -14,7 +14,8 @@ const DEFAULT_CONFIG = {
   loglevel: 'info',
 };
 
-let config = DEFAULT_CONFIG;
+let config = DEFAULT_CONFIG,
+    sentinelLib = require('@medic/sentinel');
 
 const loadTranslations = () => {
   const options = {
@@ -66,6 +67,7 @@ const initConfig = () => {
         config.schedule_evening_hours,
         config.schedule_evening_minutes
       );
+      sentinelLib.init(db, config, translations, logger);
       require('./transitions').loadTransitions();
     })
     .catch(err => {
