+++ sentinel/src/schedule/due_tasks.js
@@ -3,8 +3,9 @@
 const async = require('async'),
   _ = require('underscore'),
   moment = require('moment'),
-  utils = require('../lib/utils'),
-  date = require('../date'),
+  sentinelLib = require('@medic/sentinel'),
+  utils = sentinelLib.utils,
+  date = sentinelLib.date,
   config = require('../config'),
   db = require('../db'),
   lineage = require('@medic/lineage')(Promise, db.medic),
