--- sentinel/src/transitions/generate_patient_id_on_people.js
@@ -1,15 +1,0 @@
-var transitionUtils = require('./utils');
-
-module.exports = {
-  filter: doc => doc.type === 'person' && !doc.patient_id,
-  onMatch: change => {
-    return new Promise((resolve, reject) => {
-      transitionUtils.addUniqueId(change.doc, err => {
-        if (err) {
-          return reject(err);
-        }
-        resolve(true);
-      });
-    });
-  }
-};
